test(CountDown): add demo test
